Core Mechanics
3 things worked on today:
Creation of a nodes system, which facilitate logic and serve at the game main logi. Value is assigned from the topology, as well of the richness and the type of resource it have in 

Basic RTS Camera. Nuff' said. Pan, zoom and Drag.

Resource generations: Create resources ( wood for now) that visually show in unity. That way we we'll have some gameplay tomorrow to have some woods gatherer for prototyping