Army and Town Build
Building mode: Allow or disallow the creation of certain building depening of the intersection and rule of each building.

UI: Getting there... Pretty and more informative

Army and units: units can merge together to become an army. An army is a agglomeration of units yielding unprecedented force. They are , however, slower and required maintenance or there will be insubordination  /defect

Experimental : Mesh of certain section of the terrain, that way it'll be cooler looking to divided each nodes