@Nana : Make RTS Great Again && Defense/Economy 1.01
Some Design about economy got implemented. There is taxes on each house that the player passively receive. In the future, he will be able to select the amount of % tax which will change or not the satisfaction of the population.

The formulae for growth implemented, through generational chance and immigration. Immigration is sparse and few between since we are in the 800's century or so, but depending on the  Wealth, Housing and Security of the nation, the population will be replace. So if a player go the length to carter to these need, he'll 1.5x or 2x his population ( we need 2x to maintain population level, and 3 for 1.5x);

it's possible to upgrade housing and unit

Possible to remove building