@Nana Interactivity 1.3
Formation and Army: There is the posibility to create Army just by dragging - they act as one but with the force of many!

Drag function : Now you can carry out commands in bulk

Optimization 1: The game have some performance issue cause by water  and the nodes in editor. We'll have to decrease the resolution of nodes and search a method to make water beautiful AND optimized. 
Occlusion culling is already in use. 

AI: Smart Unit - they can detect more enemy. 
Structure: Various functions add and changed to accommodate a freeform structure. Selection is no longer a single entity, but an array of entity. In the future, we'll change it to unit, as selecting unit is useless. 
UI: indicator to know which unit we select